+++ app/controllers/wishes_controller.rb
@@ -6,16 +6,16 @@ class WishesController < ApplicationController
   def new
     @list = List.find(params[:list_id])
     @wish = Wish.new
+    authorize @wish
   end
 
   def create
     @wish = Wish.new
-    @wish.title = params[:wish][:title]
-    @wish.body = params[:wish][:body]
+    @wish.assign_attributes(wish_params)
     @list = List.find(params[:list_id])
     @wish.user = current_user
-
     @wish.list = @list
+    authorize @wish
 
     if @wish.save
      flash[:notice] = "Wish was saved successfully."
@@ -28,16 +28,17 @@ class WishesController < ApplicationController
 
   def edit
     @wish = Wish.find(params[:id])
+    authorize @wish
   end
 
   def update
     @wish = Wish.find(params[:id])
-    @wish.title = params[:wish][:title]
-    @wish.body = params[:wish][:body]
+    @wish.assign_attributes(wish_params)
+    authorize @wish
 
     if @wish.save
       flash[:notice] = "Wish was updated successfully."
-      redirect_to [@wish.list, @wish]
+      redirect_to @wish.list
     else
       flash.now[:alert] = "There was an error saving the wish. Please try again."
       render :edit
@@ -46,6 +47,7 @@ class WishesController < ApplicationController
 
   def destroy
     @wish = Wish.find(params[:id])
+    authorize @wish
 
     if @wish.destroy
       flash[:notice] = "\"#{@wish.title}\" was deleted successfully."
@@ -55,4 +57,11 @@ class WishesController < ApplicationController
       render :show
     end
   end
+
+  private
+
+  def wish_params
+    params.require(:wish).permit(:title, :body)
+  end
+
 end
